python/typing-speed-game/requirements.txt
# Core Python requirements
pillow>=9.0.0
